requirements.txt
numpy
torch
torch-scatter
torch-geometric
scikit-learn
meshio
matplotlib
PyYAML
wandb
